[INPLACE-175] 헤더 로직 분리
