+++ frontend/src/components/common/layouts/MainLayout.tsx
@@ -4,8 +4,8 @@ import styled from 'styled-components';
 import { QueryErrorResetBoundary } from '@tanstack/react-query';
 import { ErrorBoundary } from 'react-error-boundary';
 import Footer from '@/components/common/layouts/Footer';
-import Header from '@/components/common/layouts/Header';
 import Loading from '@/components/common/layouts/Loading';
+import Header from '@/components/common/layouts/Header/index';
 import MainSkeleton from '@/components/Main/MainSkeleton';
 import DetailSkeleton from '@/components/Detail/DetailSkeleton';
 import useScrollToTop from '@/hooks/useScrollToTop';
